refactor(index): separate common FRI reading from index consumers
