+++ rust/lance-index/src/scalar/btree.rs
@@ -2,7 +2,6 @@
 // SPDX-FileCopyrightText: Copyright The Lance Authors
 
 use lance_core::utils::row_addr_remap::RowAddrRemap;
-use lance_index_core::remapping::RowIdRemapping;
 use std::{
     any::Any,
     cmp::Ordering,
@@ -1418,7 +1417,7 @@ impl BTreeIndexState {
         &self,
         store: Arc<dyn IndexStore>,
         index_cache: &LanceCache,
-        frag_reuse_index: Option<RowIdRemapping>,
+        frag_reuse_index: Option<Arc<dyn RowIdRemapper>>,
     ) -> Result<Arc<dyn ScalarIndex>> {
         let index = BTreeIndex::try_from_serialized(
             self.lookup_batch.clone(),
@@ -1552,7 +1551,7 @@ pub struct BTreeIndex {
     /// - The local page_idx is calculated: `142 - 100 = 42`.
     /// - The system now knows to read page `42` from the file `part_2_page_file.lance`.
     ranges_to_files: Option<Arc<RangeInclusiveMap<u32, (String, u32)>>>,
-    frag_reuse_index: Option<RowIdRemapping>,
+    frag_reuse_index: Option<Arc<dyn RowIdRemapper>>,
 }
 
 impl DeepSizeOf for BTreeIndex {
@@ -1573,7 +1572,7 @@ impl BTreeIndex {
         index_cache: WeakLanceCache,
         batch_size: u64,
         ranges_to_files: Option<Arc<RangeInclusiveMap<u32, (String, u32)>>>,
-        frag_reuse_index: Option<RowIdRemapping>,
+        frag_reuse_index: Option<Arc<dyn RowIdRemapper>>,
     ) -> Self {
         Self {
             page_lookup,
@@ -1679,9 +1678,8 @@ impl BTreeIndex {
             .read_record_batch(page_number as u64, self.batch_size)
             .await?;
         if let Some(frag_reuse_index_ref) = self.frag_reuse_index.as_ref() {
-            serialized_page = frag_reuse_index_ref
-                .remap_row_ids_record_batch(serialized_page, 1)
-                .await?;
+            serialized_page =
+                frag_reuse_index_ref.remap_row_ids_record_batch(serialized_page, 1)?;
         }
         FlatIndex::try_new(serialized_page)
     }
@@ -1738,7 +1736,7 @@ impl BTreeIndex {
         index_cache: &LanceCache,
         batch_size: u64,
         ranges_to_files: Option<Arc<RangeInclusiveMap<u32, (String, u32)>>>,
-        frag_reuse_index: Option<RowIdRemapping>,
+        frag_reuse_index: Option<Arc<dyn RowIdRemapper>>,
     ) -> Result<Self> {
         let data_type = data.column(0).data_type().clone();
         let page_lookup = Arc::new(BTreeLookup::try_new(data)?);
@@ -1756,7 +1754,7 @@ impl BTreeIndex {
 
     async fn load(
         store: Arc<dyn IndexStore>,
-        frag_reuse_index: Option<RowIdRemapping>,
+        frag_reuse_index: Option<Arc<dyn RowIdRemapper>>,
         index_cache: &LanceCache,
     ) -> Result<Arc<Self>> {
         let (page_lookup_file, standalone_partition_page_file) =
@@ -1992,12 +1990,12 @@ fn filter_keeps_nothing(filter: &Option<OldIndexDataFilter>) -> bool {
 
 fn remap_row_ids(
     stream: SendableRecordBatchStream,
-    frag_reuse_index: RowIdRemapping,
+    frag_reuse_index: Arc<dyn RowIdRemapper>,
 ) -> SendableRecordBatchStream {
     let schema = stream.schema();
-    let remapped = stream.and_then(move |batch| {
-        let remapper = frag_reuse_index.clone();
-        async move { Ok(remapper.remap_row_ids_record_batch(batch, 1).await?) }
+    let remapped = stream.map(move |batch_result| {
+        let batch = batch_result?;
+        Ok(frag_reuse_index.remap_row_ids_record_batch(batch, 1)?)
     });
     Box::pin(RecordBatchStreamAdapter::new(schema, remapped))
 }
@@ -3360,26 +3358,7 @@ impl ScalarIndexPlugin for BTreeIndexPlugin {
         frag_reuse_index: Option<Arc<dyn RowIdRemapper>>,
         cache: &LanceCache,
     ) -> Result<Arc<dyn ScalarIndex>> {
-        Ok(BTreeIndex::load(
-            index_store,
-            frag_reuse_index.map(RowIdRemapping::InMemory),
-            cache,
-        )
-        .await? as Arc<dyn ScalarIndex>)
-    }
-
-    fn supports_batch_row_id_remapping(&self) -> bool {
-        true
-    }
-
-    async fn load_index_with_remapping(
-        &self,
-        index_store: Arc<dyn IndexStore>,
-        _index_details: &prost_types::Any,
-        remapping: Option<RowIdRemapping>,
-        cache: &LanceCache,
-    ) -> Result<Arc<dyn ScalarIndex>> {
-        Ok(BTreeIndex::load(index_store, remapping, cache).await?)
+        Ok(BTreeIndex::load(index_store, frag_reuse_index, cache).await? as Arc<dyn ScalarIndex>)
     }
 
     async fn get_from_cache(
@@ -3394,7 +3373,7 @@ impl ScalarIndexPlugin for BTreeIndexPlugin {
         Ok(Some(state.reconstruct(
             index_store,
             cache,
-            frag_reuse_index.map(RowIdRemapping::InMemory),
+            frag_reuse_index,
         )?))
     }
 
@@ -3423,13 +3402,7 @@ impl ScalarIndexPlugin for BTreeIndexPlugin {
             BTreeIndexStateKey,
             load,
             BTreeIndexState::from_index,
-            move |state| {
-                state.reconstruct(
-                    index_store,
-                    cache,
-                    frag_reuse_index.map(RowIdRemapping::InMemory),
-                )
-            },
+            move |state| state.reconstruct(index_store, cache, frag_reuse_index),
         )
         .await
     }
@@ -3438,7 +3411,6 @@ impl ScalarIndexPlugin for BTreeIndexPlugin {
 #[cfg(test)]
 mod tests {
     use lance_core::utils::row_addr_remap::RowAddrRemap;
-    use lance_index_core::remapping::RowIdRemapping;
     use std::sync::atomic::Ordering;
     use std::{collections::HashMap, sync::Arc};
 
@@ -6522,7 +6494,7 @@ mod tests {
             .reconstruct(
                 test_store.clone(),
                 &LanceCache::no_cache(),
-                Some(RowIdRemapping::InMemory(frag_reuse_index)),
+                Some(frag_reuse_index),
             )
             .unwrap();
