+++ rust/lance-index/src/scalar/inverted/index/doc_set.rs
@@ -410,30 +410,7 @@ impl DocSet {
         is_legacy: bool,
         frag_reuse_index: Option<Arc<dyn RowIdRemapper>>,
     ) -> Result<Self> {
-        Self::load_with_remapping(
-            reader,
-            is_legacy,
-            frag_reuse_index.map(RowIdRemapping::InMemory),
-        )
-        .await
-    }
-
-    pub(crate) async fn load_with_remapping(
-        reader: Arc<dyn IndexReader>,
-        is_legacy: bool,
-        remapping: Option<RowIdRemapping>,
-    ) -> Result<Self> {
         let batch = reader.read_range(0..reader.num_rows(), None).await?;
-        let (batch, frag_reuse_index) = match remapping {
-            Some(remapping) => {
-                let row_id_idx = batch.schema().index_of(ROW_ID)?;
-                let (batch, remapper) = remapping
-                    .remap_row_ids_preserving_layout(batch, row_id_idx)
-                    .await?;
-                (batch, Some(remapper))
-            }
-            None => (batch, None),
-        };
         let row_id_col = batch[ROW_ID].as_primitive::<datatypes::UInt64Type>();
         let num_tokens_col = batch[NUM_TOKEN_COL].as_primitive::<datatypes::UInt32Type>();
         let mut doc_indices = Vec::new();
