+++ rust/lance/src/index.rs
@@ -80,7 +80,7 @@ mod api;
 pub(crate) mod append;
 mod create;
 pub mod frag_reuse;
-pub(crate) mod frag_reuse_query;
+pub mod frag_reuse_reader;
 pub mod mem_wal;
 pub mod prefilter;
 pub mod scalar;
@@ -1878,7 +1878,7 @@ impl DatasetIndexExt for Dataset {
             .iter()
             .find(|idx| idx.name == FRAG_REUSE_INDEX_NAME && idx.index_version != 0)
         {
-            let mapping = frag_reuse_query::QueryFragReuseIndex::open(self, fri).await?;
+            let mapping = frag_reuse_reader::FragmentReuseIndex::open(self, fri).await?;
             let mut supported = Vec::with_capacity(indices.len());
             for index in indices.iter() {
                 if !index_is_usable(index) {
@@ -1887,22 +1887,9 @@ impl DatasetIndexExt for Dataset {
                 if index.name != FRAG_REUSE_INDEX_NAME
                     && mapping.may_need_translation(index.fragment_bitmap.as_ref())
                 {
-                    let can_remap = if segment_has_vector_details(index) {
-                        frag_reuse_query::vector_supports_batch_remapping(self, index).await?
-                    } else {
-                        index
-                            .index_details
-                            .as_ref()
-                            .and_then(|details| {
-                                scalar::SCALAR_INDEX_PLUGIN_REGISTRY
-                                    .get_plugin_by_details(details)
-                                    .ok()
-                            })
-                            .is_some_and(|plugin| plugin.supports_batch_row_id_remapping())
-                    };
-                    if index.fragment_bitmap.is_none() || !can_remap {
-                        continue;
-                    }
+                    // The consumer integration is a separate layer. Until then,
+                    // affected segments must scan even when coverage is derivable.
+                    continue;
                 }
                 supported.push(index.clone());
             }
@@ -3086,52 +3073,33 @@ impl DatasetIndexInternalExt for Dataset {
             .await?
             .ok_or_else(|| Error::index(format!("Index with id {} does not exist", uuid)))?;
         let object_store = self.object_store_for_index(&index_meta).await?;
-        let resolved = frag_reuse::open_row_id_remapping(self, &index_meta, metrics).await?;
-        let remapping = resolved.map(|(_, remapping)| remapping);
-        let is_tagged = matches!(
-            remapping,
-            Some(lance_index::scalar::RowIdRemapping::External(_))
-        );
-        let query_cache = crate::session::index_caches::DSIndexCache(if is_tagged {
-            self.index_cache
-                .with_key_prefix(self.manifest_location.path.as_ref())
-        } else {
-            self.index_cache.0.clone()
-        });
 
         // Check sized cache first (v2+ indices with serializable state).
         let state_key = IvfIndexStateCacheKey::new(uuid, frag_reuse_uuid.as_ref());
-        if let Some(entry) = query_cache.get_with_key(&state_key).await {
+        if let Some(entry) = self.index_cache.get_with_key(&state_key).await {
             log::debug!("Found IvfIndexState in cache uuid: {}", uuid);
-            let partition_cache = query_cache.for_index(uuid, frag_reuse_uuid.as_ref());
+            let partition_cache = self.index_cache.for_index(uuid, frag_reuse_uuid.as_ref());
+            let frag_reuse_index =
+                frag_reuse::open_legacy_query_remapper(self, &index_meta, metrics).await?;
             return entry
                 .0
                 .reconstruct(
                     object_store,
                     self.metadata_cache.as_ref(),
                     partition_cache,
-                    remapping,
+                    frag_reuse_index,
                 )
                 .await;
         }
 
         // Fallback: in-memory cache for legacy indices.
         let cache_key = LegacyVectorIndexCacheKey::new(uuid, frag_reuse_uuid.as_ref());
-        if let Some(cached) = query_cache.get_with_key(&cache_key).await {
+        if let Some(cached) = self.index_cache.get_with_key(&cache_key).await {
             return Ok(cached.0.clone());
         }
 
-        // Only legacy vector file readers consume the V1 handle. A tagged
-        // history is handled by the shared remapper above, including identity.
-        let has_tagged_history = load_all_indices(self)
-            .await?
-            .iter()
-            .any(|index| index.name == FRAG_REUSE_INDEX_NAME && index.index_version != 0);
-        let frag_reuse_index = if has_tagged_history {
-            None
-        } else {
-            self.open_frag_reuse_index(metrics).await?
-        };
+        let frag_reuse_index =
+            frag_reuse::open_legacy_query_remapper(self, &index_meta, metrics).await?;
         let index_dir = self.indice_files_dir(&index_meta)?;
         let index_file = index_dir
             .clone()
@@ -3153,7 +3121,7 @@ impl DatasetIndexInternalExt for Dataset {
         // Namespace the index cache by the UUID of the index. v2+ partition
         // entries are store-free and remain reusable across object-store
         // generations alongside their serializable state.
-        let index_cache = query_cache.for_index(uuid, frag_reuse_uuid.as_ref());
+        let index_cache = self.index_cache.for_index(uuid, frag_reuse_uuid.as_ref());
 
         // Extract the cacheable state before type-erasing to Arc<dyn VectorIndex>.
         fn wrap_ivf<S: IvfSubIndex + 'static, Q: Quantization + 'static>(
@@ -3249,7 +3217,7 @@ impl DatasetIndexInternalExt for Dataset {
                                 object_store.clone(),
                                 index_dir,
                                 uuid.to_owned(),
-                                remapping,
+                                frag_reuse_index,
                                 self.metadata_cache.as_ref(),
                                 index_cache,
                                 file_sizes,
@@ -3262,7 +3230,7 @@ impl DatasetIndexInternalExt for Dataset {
                                 object_store.clone(),
                                 index_dir,
                                 uuid.to_owned(),
-                                remapping,
+                                frag_reuse_index,
                                 self.metadata_cache.as_ref(),
                                 index_cache,
                                 file_sizes,
@@ -3281,7 +3249,7 @@ impl DatasetIndexInternalExt for Dataset {
                             object_store.clone(),
                             index_dir,
                             uuid.to_owned(),
-                            remapping,
+                            frag_reuse_index,
                             self.metadata_cache.as_ref(),
                             index_cache,
                             file_sizes,
@@ -3295,7 +3263,7 @@ impl DatasetIndexInternalExt for Dataset {
                             object_store.clone(),
                             index_dir,
                             uuid.to_owned(),
-                            remapping,
+                            frag_reuse_index,
                             self.metadata_cache.as_ref(),
                             index_cache,
                             file_sizes,
@@ -3309,7 +3277,7 @@ impl DatasetIndexInternalExt for Dataset {
                             object_store.clone(),
                             index_dir,
                             uuid.to_owned(),
-                            remapping,
+                            frag_reuse_index,
                             self.metadata_cache.as_ref(),
                             index_cache,
                             file_sizes,
@@ -3324,7 +3292,7 @@ impl DatasetIndexInternalExt for Dataset {
                                 object_store.clone(),
                                 index_dir,
                                 uuid.to_owned(),
-                                remapping,
+                                frag_reuse_index,
                                 self.metadata_cache.as_ref(),
                                 index_cache,
                                 file_sizes,
@@ -3337,7 +3305,7 @@ impl DatasetIndexInternalExt for Dataset {
                                 object_store.clone(),
                                 index_dir,
                                 uuid.to_owned(),
-                                remapping,
+                                frag_reuse_index,
                                 self.metadata_cache.as_ref(),
                                 index_cache,
                                 file_sizes,
@@ -3352,7 +3320,7 @@ impl DatasetIndexInternalExt for Dataset {
                             object_store.clone(),
                             index_dir,
                             uuid.to_owned(),
-                            remapping,
+                            frag_reuse_index,
                             self.metadata_cache.as_ref(),
                             index_cache,
                             file_sizes,
@@ -3366,7 +3334,7 @@ impl DatasetIndexInternalExt for Dataset {
                             object_store.clone(),
                             index_dir,
                             uuid.to_owned(),
-                            remapping,
+                            frag_reuse_index,
                             self.metadata_cache.as_ref(),
                             index_cache,
                             file_sizes,
@@ -3398,11 +3366,11 @@ impl DatasetIndexInternalExt for Dataset {
             io_stats.add_scan_stats(&open_stats);
         }
         if let Some(ivf_entry) = ivf_entry {
-            query_cache
+            self.index_cache
                 .insert_with_key(&state_key, Arc::new(ivf_entry))
                 .await;
         } else {
-            query_cache
+            self.index_cache
                 .insert_with_key(&cache_key, Arc::new(CachedLegacyVectorIndex(index.clone())))
                 .await;
         }
