+++ rust/lance/src/index/scalar.rs
@@ -16,7 +16,6 @@ pub(crate) mod rtree;
 pub(crate) mod zonemap;
 
 pub use inverted::{load_segment_details, load_segment_params, load_segments};
-use lance_index::scalar::RowIdRemapping;
 
 pub use crate::index::scalar_logical::{LogicalScalarIndex, load_named_scalar_segments};
 
@@ -40,6 +39,7 @@ use lance_core::datatypes::Field;
 use lance_core::utils::tracing::{IO_TYPE_OPEN_SCALAR, TRACE_IO_EVENTS};
 use lance_core::{Error, ROW_ADDR, ROW_ID, Result};
 use lance_datafusion::exec::LanceExecutionOptions;
+use lance_index::frag_reuse::CompactFragReuseIndexHandle;
 use lance_index::metrics::{MetricsCollector, NoOpMetricsCollector};
 use lance_index::pb::VectorIndexDetails;
 use lance_index::pbold::{
@@ -58,8 +58,8 @@ use lance_index::scalar::registry::{
 };
 use lance_index::scalar::{BuiltinIndexType, CreatedIndex, InvertedIndexParams};
 use lance_index::scalar::{
-    ScalarIndex, ScalarIndexParams, bitmap::BITMAP_LOOKUP_NAME, inverted::INVERT_LIST_FILE,
-    lance_format::LanceIndexStore,
+    RowIdRemapper, ScalarIndex, ScalarIndexParams, bitmap::BITMAP_LOOKUP_NAME,
+    inverted::INVERT_LIST_FILE, lance_format::LanceIndexStore,
 };
 use lance_index::{IndexCriteria, IndexType};
 use lance_table::format::{Fragment, IndexMetadata};
@@ -283,7 +283,7 @@ pub(crate) async fn load_fts_training_data(
 }
 
 // TODO: Allow users to register their own plugins
-pub(crate) static SCALAR_INDEX_PLUGIN_REGISTRY: LazyLock<Arc<IndexPluginRegistry>> =
+static SCALAR_INDEX_PLUGIN_REGISTRY: LazyLock<Arc<IndexPluginRegistry>> =
     LazyLock::new(IndexPluginRegistry::with_default_plugins);
 
 pub struct IndexDetails(pub Arc<prost_types::Any>);
@@ -527,7 +527,7 @@ async fn validate_label_list_index_compatibility(
     dataset: &Dataset,
     column: &str,
     index: &IndexMetadata,
-    index_store: &dyn IndexStore,
+    index_store: &Arc<LanceIndexStore>,
 ) -> Result<()> {
     let Some(field) = dataset.schema().field(column) else {
         return Ok(());
@@ -572,41 +572,31 @@ pub async fn open_scalar_index(
     let index_details = fetch_index_details(dataset, column, index).await?;
     let plugin = SCALAR_INDEX_PLUGIN_REGISTRY.get_plugin_by_details(index_details.as_ref())?;
 
-    let resolved = super::frag_reuse::open_row_id_remapping(dataset, index, metrics).await?;
-    let cache_id = resolved.as_ref().map(|(uuid, _)| uuid);
-    let remapping = resolved.as_ref().map(|(_, remapping)| remapping.clone());
-    let is_tagged = matches!(remapping, Some(RowIdRemapping::External(_)));
-    let index_cache = dataset.index_cache.for_index(&index.uuid, cache_id);
-    let index_cache = if is_tagged {
-        index_cache.with_key_prefix(dataset.manifest_location.path.as_ref())
-    } else {
-        index_cache
-    };
-    let frag_reuse_index = match &remapping {
-        Some(RowIdRemapping::InMemory(remapper)) => Some(remapper.clone()),
-        _ => None,
-    };
+    let frag_reuse_index =
+        super::frag_reuse::open_legacy_query_remapper(dataset, index, metrics).await?;
+
+    let index_cache = dataset
+        .index_cache
+        .for_index(&index.uuid, frag_reuse_index.as_ref().map(|f| &f.uuid));
+
+    let frag_reuse_index: Option<Arc<dyn RowIdRemapper>> = frag_reuse_index
+        .map(|f| Arc::new(CompactFragReuseIndexHandle(f)) as Arc<dyn RowIdRemapper>);
 
     // Runs only on a cold miss, and at most once even under concurrent opens
     // (the plugin coalesces). The compat check lives here because a warm hit was
     // already validated this session, saving the extra `open_index_file` IOP.
     let load: ScalarIndexLoad = Box::pin({
         let index_store = index_store.clone();
-        let remapping = remapping.clone();
+        let frag_reuse_index = frag_reuse_index.clone();
         let index_cache = index_cache.clone();
         async move {
             if index_details.type_url.ends_with("LabelListIndexDetails") {
-                validate_label_list_index_compatibility(
-                    dataset,
-                    column,
-                    index,
-                    index_store.as_ref(),
-                )
-                .await?;
+                validate_label_list_index_compatibility(dataset, column, index, &index_store)
+                    .await?;
             }
 
             let index = plugin
-                .load_index_with_remapping(index_store, &index_details, remapping, &index_cache)
+                .load_index(index_store, &index_details, frag_reuse_index, &index_cache)
                 .await?;
 
             tracing::info!(target: TRACE_IO_EVENTS, index_uuid = %index_uuid, r#type = IO_TYPE_OPEN_SCALAR, index_type = index.index_type().to_string());
@@ -615,36 +605,19 @@ pub async fn open_scalar_index(
         }
     });
 
-    if is_tagged {
-        lance_index::scalar::registry::single_flight_store_bound_open(
-            index_store,
-            &index_cache,
-            load,
-        )
+    plugin
+        .get_or_insert_in_cache(index_store, frag_reuse_index, &index_cache, load)
         .await
-    } else {
-        plugin
-            .get_or_insert_in_cache(index_store, frag_reuse_index, &index_cache, load)
-            .await
-    }
 }
 
 pub(crate) async fn cached_scalar_index_container(
     dataset: &Dataset,
     uuid: &Uuid,
 ) -> Option<Arc<dyn ScalarIndex>> {
-    let metadata = dataset.load_indices().await.ok()?;
-    let fri = metadata
-        .iter()
-        .find(|idx| idx.name == lance_index::frag_reuse::FRAG_REUSE_INDEX_NAME);
+    let frag_reuse_uuid = dataset.frag_reuse_index_uuid().await;
     let index_cache = dataset
         .index_cache
-        .for_index(uuid, fri.map(|idx| &idx.uuid));
-    let index_cache = if fri.is_some_and(|idx| idx.index_version != 0) {
-        index_cache.with_key_prefix(dataset.manifest_location.path.as_ref())
-    } else {
-        index_cache
-    };
+        .for_index(uuid, frag_reuse_uuid.as_ref());
     index_cache
         .get_unsized_with_key(&ScalarIndexCacheKey)
         .await
